refactor: group uv candidate runtime
